data: qmi-service-ssc: add SSC service
The SSC service is the Qualcomm Snapdragon Sensor Core (SSC) which is a
remoteproc exposing the sensor devices in the latest Qualcomm SoCs.
The SSC manages the power to the sensors, sample rates, etc. which
can be used over QMI. In contrast to other QMI services, the SSC uses
QMI only as a transport mechamism. The actual messages are Protobuf
messages encoded in the QMI messages and indications.
Therefore, only 1 QMI message is used and 2 QMI indications.
The SSC may emit QMI indications with a small data size or much bigger,
they are indentical besides the data size and ID.