uim: new UIM personalization feature enum in Card Status messages
The personalization feature enum in "UIM Card Status" QMI message is
treated differently in QMI modems, and is not same as the existing
'QmiUimCardApplicationPersonalizationFeature' enum. A new enum
'QmiUimCardApplicationPersonalizationFeatureStatus' for the "UIM Card
Status" QMI message is introduced.

The enum change in the QMI message leads to the need of compat symbols
to support API/ABI compatibility. Given that the format and size of
the struct involved in the enum change is the same, i.e. only the
representation of the integer as a different enum changed, we don't
treat the change as a whole TLV rename. Instead, we only change the
struct type that is stored in the GArray/GPtrArray.

We also provide a compat symbol for the UNKNOWN enum name, which the
old enum no longer has.

Originally written by Akash Aggarwal <quic_akasagga@quicinc.com> in
5f366be3000e8c00f7ce73d64606095d3ba4c22e, but later reverted to handle
all the needed compat symbols.