qmicli,atr: add monitoring unsolicited action
Add an action to watch for unsolicited notifications.

To enable the AT communication in the QMI client and get the unsolicited
notifications, it is required to send any initial AT command.

The implementation uses 'ATE0' so that the echo is disabled in the
monitoring client: this could also be useful if the same client cid
is used also for AT commands sending through --atr-send-only to avoid
polluting the monitoring output.