libqmi-glib,message: fix 16-bit service on big endian architectures
The latest release introduces handling of 16-bit service indications.
However, only the raw message data is returned from
`qmi_message_get_service()`, leading to incorrect values on big-endian
architectures. As a consequence, `libqmi` ultimately fails to build as
the corresponding test errors out in this case.

This patch ensures the service indication is correct on all kinds of
architectures.