mbim-codegen,struct: provide expected struct size to struct reader method
If the struct size is known in advance, e.g. if the struct is of fixed
size, or if the struct is given via an OL pair (so 'L' is the size of
the struct), pass down the size information to the struct reader
method, as it may be able to perform additional checks based on that.