libmbim-glib,device: avoid emitting SIGNAL_ERROR if device closed
The SIGNAL_ERROR signal is supposed to be emitted whenever we receive
a FUNCTION_ERROR response. The signal is emitted *after* the message
response processor has been executed, and this means that there is
a chance that the response processor may have already closed the MBIM
device. If this happens, we should not emit the SIGNAL_ERROR signal,
as the user has already expressed their desire to fully close the
ongoing MBIM session.

This avoids the SIGNAL_ERROR signal being processed after the user has
already requested a new MBIM device open following a forced close.

Issue found by Maciej S. Szmigiero <mail@maciej.szmigiero.name>, and
this patch implemented as an alternative to the one he wrote in
https://gitlab.freedesktop.org/mobile-broadband/libmbim/-/merge_requests/193

Fixes #43