libmbim-glib,proxy: no need to block MBIM_DEVICE_SIGNAL_ERROR during open
If the proxy is checking whether a given device is really open or not,
it sends a caps query request, which may be replied with a
FUNCTION_ERROR stating that the device is not open. If this happens,
the proxy will explicitly force close the device, and try to reopen it
again.

The MBIM_DEVICE_SIGNAL_ERROR is always scheduled to be emitted after
the response processor has been executed. In the case of the proxy,
the response processor force-closes the MBIM device, which in turn
would cancel any scheduled MBIM_DEVICE_SIGNAL_ERROR emission.

There is really no need to block MBIM_DEVICE_SIGNAL_ERROR any more due
to this.