feat: add optional key-value-pair parameters to allow specification of a session-type
This commit makes it possible to unlock other SIM cards in multi-SIM environments.
Instead of using the hard-coded session-type `card-slot-1` it is now possible to append `,session-type=card-slot-2` to the parameters of the following actions:

- `--uim-set-pin-protection=[(PIN1|PIN2|UPIN),(disable|enable),(current PIN)[,"key=value,..."]]`
- `--uim-verify-pin=[(PIN1|PIN2|UPIN),(current PIN)[,"key=value,..."]]`
- `--uim-unblock-pin=[(PIN1|PIN2|UPIN),(PUK),(new PIN)[,"key=value,..."]]`
- `--uim-change-pin=[(PIN1|PIN2|UPIN),(old PIN),(new PIN)[,"key=value,..."]]`

This change is backward compatible because the key-value-pair parameters are optional and without them `card-slot-1` will be used.