data: qmi-service-imsa: add indications
Indications on the IMSA service allow the Application Processor to be
notified when the modem's IMS state changes. Add the indications, which
are similar to Get IMS Services/Registration Status responses. Since the
Application Processor must subscribe to the indications to receive them,
also add the IMSA Register Indications request.

There is another 0x33 indication which corresponds to field 0x1B of the
IMSA Register Indications request, but omit it until it is understood
enough to be named.