mbimcli-intel-tools: parse trace command & value in a more straightforward way
Don't conditionalize setting trace_command and trace_value on split
substrings being non-NULL. It makes a static analysis tool think they
might be used uninitialized. That can-not happen, because the substrings
can in fact never be NULL.

Let's keep the check in form of an assert (perhaps to guard against
a possible glib bug, etc.).