mbimcli: new '--set-register-state' action
The new '--set-register-state' action allows configuring automatic or
manual registration, also specifying the MCCMNC to bind to and
optionally the list of data modes to prefer.

The '--register-automatic' action is considered deprecated, and can be
replaced with '--set-register-state="action=automatic".

Automatic registration setup:
  $ sudo mbimcli -d /dev/cdc-wdm1 -p --set-register-state='
               action=automatic'

Manual registration to a specific operator:
  $ sudo mbimcli -d /dev/cdc-wdm1 -p --set-register-state='
               action=manual,
               provider-id=21407'

Manual registration to a specific operator on limited access techs:
  $ sudo mbimcli -d /dev/cdc-wdm1 -p --set-register-state='
               action=manual,
               provider-id=21407,
               data-class=lte|umts'